Can now read and write to the savedata files
Try running the server once to initialize the ArenaSaveData.ini file. Then open it and you can edit the values, though the only thing it changes right now is the SpawnLocationArray, in which all these values are stored right now. None of this code should affect current operations stuff. Saving the data happens using SaveArena() in OnDisable() in main.lua.